Add support for PE binaries
This adds support for listing sections in PE binaries. BSS is calculated
by taking the difference of on disk size of the data section and the
virtual size. The section name type needed to be switch to a String due
to the &str ref going out of scope for the PE variant.